+++ modules/nixos/performance/zram.nix
@@ -13,6 +13,11 @@ in {
 
   config = mkIf cfg.enable {
     # Enable ZRAM swap for better performance
-    zramSwap.enable = true;
+    zramSwap = {
+      enable = true;
+      algorithm = "lz4";
+      memoryPercent = 100;
+      priority = 999;
+    };
   };
 }